+++ vrp-core/src/solver/evolution/config.rs
@@ -6,6 +6,8 @@ use crate::solver::evolution::{EvolutionStrategy, RunSimple};
 use crate::solver::hyper::{HyperHeuristic, MultiSelective};
 use crate::solver::mutation::*;
 use crate::solver::population::*;
+use crate::solver::processing::post::{AdvanceDeparture, PostProcessing};
+use crate::solver::processing::pre::PreProcessing;
 use crate::solver::telemetry::Telemetry;
 use crate::solver::termination::*;
 use crate::solver::TelemetryMode;
@@ -17,7 +19,10 @@ pub struct EvolutionConfig {
     /// An original problem.
     pub problem: Arc<Problem>,
 
-    /// A population configuration
+    /// A pre/post processing configuration.
+    pub processing: ProcessingConfig,
+
+    /// A population configuration.
     pub population: PopulationConfig,
 
     /// A hyper heuristic.
@@ -48,6 +53,14 @@ pub struct PopulationConfig {
     pub variation: Option<Box<dyn Population + Send + Sync>>,
 }
 
+/// A configuration which keeps track of pre/post processing settings.
+pub struct ProcessingConfig {
+    /// Preprocessing settings.
+    pub pre: Option<Arc<dyn PreProcessing + Send + Sync>>,
+    /// Postprocessing settings.
+    pub post: Option<Arc<dyn PostProcessing + Send + Sync>>,
+}
+
 /// An initial solutions configuration.
 pub struct InitialConfig {
     /// Create methods to produce initial individuals.
@@ -65,6 +78,7 @@ impl EvolutionConfig {
     pub fn new(problem: Arc<Problem>, environment: Arc<Environment>) -> Self {
         Self {
             problem: problem.clone(),
+            processing: ProcessingConfig { pre: None, post: Some(Arc::new(AdvanceDeparture::default())) },
             population: PopulationConfig {
                 initial: InitialConfig {
                     max_size: 7,