+++ vrp-pragmatic/tests/features/breaks/interval_break_test.rs
@@ -306,10 +306,8 @@ fn can_consider_departure_rescheduling() {
     };
     let matrix = create_matrix_from_problem(&problem);
 
-    let solution = solve_with_cheapest_insertion(problem, Some(vec![matrix]));
+    let solution = solve_with_metaheuristic(problem, Some(vec![matrix]));
 
-    assert_eq!(
-        solution.violations,
-        Some(vec![Violation::Break { vehicle_id: "my_vehicle_1".to_string(), shift_index: 0 }])
-    );
+    assert!(solution.violations.is_none());
+    assert!(solution.unassigned.is_none());
 }